# api/openapi.yaml
openapi: 3.0.4
info:
  title: Authorization service
  description: Proxy  API
  version: 0.1.9

servers:
  - url: http://localhost:8083/
    description: Local development server
  - url: https://www.handmadejewelry.com/
    description: Production server

paths:
  /login:
    post:
      summary: Login user
      description: Authenticates user by login and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
      responses:
        "200":
          description: Successful authentication
        "400":
          description: Invalid data
  /logout:
    post:
      summary: Logout user
      description: Logs out user by invalidating refresh token and clearing authentication cookies.
      responses:
        "200":
          description: Successful logout
        "401":
          description: Unauthorized - invalid or missing authentication

  /refresh-token:
    get:
      summary: Refresh access token
      description: Refresh access token using a refresh token
      responses:
        "200":
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: New access token
                  refresh_token:
                    type: string
                    description: New refresh token
        "401":
          description: Unauthorized — invalid or expired refresh token
      security: [ ]
      parameters:
        - in: cookie
          name: refresh_token
          required: true
          schema:
            type: string
          description: Refresh token stored in HttpOnly cookie

  /admin/service:
    get:
      summary: Returns a list of services
      description: Returns a list of services
      responses:
        "200":
          description: Successfully retrieved list of services
    post:
      summary: Create new service
      description: Create new service
      responses:
        "200":
          description: Service successfully created
    put:
      summary: Update service
      description: Update service
      responses:
        "200":
          description: Service successfully updated
    delete:
      summary: Delete service
      description: Delete service
      responses:
        "200":
          description: Service successfully deleted